+++ extensions/lorem-ipsum/src/lists.ts
@@ -2,5 +2,5 @@ import { LaunchProps } from "@vicinae/api";
 import { runNoView } from "./lib/run";
 
 export default async function Command(props: LaunchProps<{ arguments: Arguments.Lists }>) {
-  await runNoView("list", props.arguments?.count, 5);
+  await runNoView("list", props.arguments?.count);
 }